Clean up: deduplicate test Throttlers and move generic Compression code
