+++ src/main/kotlin/org/wfanet/panelmatch/common/compression/BrotliDictionaryBuilder.kt
@@ -12,13 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.panelmatch.client.common
+package org.wfanet.panelmatch.common.compression
 
 import com.google.protobuf.ByteString
 import org.wfanet.measurement.common.flatten
-import org.wfanet.panelmatch.common.compression.CompressorFactory
-import org.wfanet.panelmatch.common.compression.Dictionary
-import org.wfanet.panelmatch.common.compression.dictionary
 
 /** [DictionaryBuilder] that uses Brotli compression (via JNI). */
 class BrotliDictionaryBuilder : DictionaryBuilder {
