+++ src/main/kotlin/org/wfanet/panelmatch/common/compression/DictionaryBuilder.kt
@@ -12,13 +12,10 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.panelmatch.client.common
+package org.wfanet.panelmatch.common.compression
 
 import com.google.protobuf.ByteString
 import java.io.Serializable
-import org.wfanet.panelmatch.common.compression.Compressor
-import org.wfanet.panelmatch.common.compression.CompressorFactory
-import org.wfanet.panelmatch.common.compression.Dictionary
 
 /**
  * [Compressor] factory that supports training.
