+++ src/main/kotlin/org/wfanet/panelmatch/common/compression/UncompressedDictionaryBuilder.kt
@@ -12,13 +12,9 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package org.wfanet.panelmatch.client.common
+package org.wfanet.panelmatch.common.compression
 
 import com.google.protobuf.ByteString
-import org.wfanet.panelmatch.common.compression.CompressorFactory
-import org.wfanet.panelmatch.common.compression.Dictionary
-import org.wfanet.panelmatch.common.compression.NoOpCompressor
-import org.wfanet.panelmatch.common.compression.dictionary
 
 /**
  * Trivial trainer for [NoOpCompressor].
