+++ src/test/kotlin/org/wfanet/panelmatch/client/eventpreprocessing/PreprocessEventsInPipelineTest.kt
@@ -20,10 +20,10 @@ import org.apache.beam.sdk.values.KV
 import org.junit.Test
 import org.junit.runner.RunWith
 import org.junit.runners.JUnit4
-import org.wfanet.panelmatch.client.common.UncompressedDictionaryBuilder
 import org.wfanet.panelmatch.client.common.testing.eventsOf
 import org.wfanet.panelmatch.common.beam.testing.BeamTestBase
 import org.wfanet.panelmatch.common.beam.testing.assertThat
+import org.wfanet.panelmatch.common.compression.UncompressedDictionaryBuilder
 import org.wfanet.panelmatch.common.compression.dictionary
 import org.wfanet.panelmatch.common.toByteString
